roachtest: delete archaic jobs/mixed-version test
This roachtest has dubious intentions. The expected logic
on job cancellation and failure is specific to the job type.
It is more useful to write job type specific unit tests
rather than a roachtest to test these behaviours.

Fixes: #67587
Release note: None
