+++ pkg/cmd/roachtest/tests/BUILD.bazel
@@ -100,7 +100,6 @@ go_library(
         "mixed_version_decl_schemachange_compat.go",
         "mixed_version_decommission.go",
         "mixed_version_job_compatibility_in_declarative_schema_changer.go",
-        "mixed_version_jobs.go",
         "mixed_version_schemachange.go",
         "multitenant.go",
         "multitenant_distsql.go",
